+++ governance/program/src/state/governance.rs
@@ -31,11 +31,11 @@ pub struct GovernanceConfig {
     /// Minimum number of tokens a governance token owner must possess to be able to create a proposal
     pub min_tokens_to_create_proposal: u16,
 
-    /// Minimum waiting time in slots for an instruction to be executed after proposal is voted on
-    pub min_instruction_hold_up_time: u64,
+    /// Minimum waiting time in seconds for an instruction to be executed after proposal is voted on
+    pub min_instruction_hold_up_time: u32,
 
-    /// Time limit in slots for proposal to be open for voting
-    pub max_voting_time: u64,
+    /// Time limit in seconds for proposal to be open for voting
+    pub max_voting_time: u32,
 }
 
 /// Governance Account
